pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seo-auditor-app"
version = "0.1.0"
description = "Production-minded SEO and performance auditing web app built with FastAPI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "beautifulsoup4>=4.12.3",
  "boto3>=1.37.0",
  "fastapi>=0.115.8",
  "httpx>=0.28.1",
  "jinja2>=3.1.5",
  "lxml>=5.3.0",
  "pandas>=2.2.3",
  "playwright>=1.50.0",
  "psycopg[binary]>=3.2.3",
  "pydantic-settings>=2.7.1",
  "python-json-logger>=3.2.1",
  "python-multipart>=0.0.20",
  "redis>=5.2.1",
  "rq>=1.16.2",
  "sqlalchemy>=2.0.37",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.34.0",
  "weasyprint>=64.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
